# Go_Server/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个集合了多种语言的Web靶场，该系统是以食谱菜单管理系统为场景去编写，一种实战化形式的安全漏洞靶场，其中存在多个安全漏洞，需要我们去探索和发现。\\n\\n该项目旨在帮助安全研究人员和爱好者了解和掌握关于不同语言系统的渗透测试和代码审计知识。",
        "title": "食谱菜单管理系统靶场",
        "contact": {
            "name": "Vulnerabilities_Server",
            "url": "https://github.com/A7cc/Vulnerabilities_Server"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/food/add": {
            "post": {
                "tags": [
                    "鉴权接口-菜品相关方法"
                ],
                "summary": "添加菜品",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "接收添加菜品表单数据",
                        "name": "AddFoodRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.AddFoodRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/food/delete/{id}": {
            "delete": {
                "tags": [
                    "鉴权接口-菜品相关方法"
                ],
                "summary": "删除菜品",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "菜品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/food/detail": {
            "get": {
                "tags": [
                    "鉴权接口-菜品相关方法"
                ],
                "summary": "获取菜品详情信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "获取菜品ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/food/get": {
            "get": {
                "tags": [
                    "鉴权接口-菜品相关方法"
                ],
                "summary": "获取菜品列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/food/update": {
            "put": {
                "tags": [
                    "鉴权接口-菜品相关方法"
                ],
                "summary": "更新菜品信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "更新菜品信息",
                        "name": "UpdateFoodRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.UpdateFoodRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/food/upfoodicon": {
            "post": {
                "tags": [
                    "鉴权接口-菜品相关方法"
                ],
                "summary": "上传菜品ICON",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "上传菜品ICON",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/food/upfoodvideo": {
            "post": {
                "tags": [
                    "鉴权接口-菜品相关方法"
                ],
                "summary": "上传菜品视频",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "上传菜品视频",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/home/get": {
            "get": {
                "tags": [
                    "鉴权接口-首页设置方法"
                ],
                "summary": "获取系统数据信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/home/getsentence": {
            "get": {
                "tags": [
                    "鉴权接口-首页设置方法"
                ],
                "summary": "获取每日金句",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "获取每日金句的url",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/home/updateInfo": {
            "put": {
                "tags": [
                    "鉴权接口-首页设置方法"
                ],
                "summary": "用户自己更新个人信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户自己更新个人信息参数",
                        "name": "UpdateInfoType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.UpdateInfoType"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/home/updatePwd": {
            "put": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "鉴权接口-首页设置方法"
                ],
                "summary": "修改个人密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户自己更新密码和UID",
                        "name": "PwdDataType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.PwdDataType"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/home/upuseravatar": {
            "post": {
                "tags": [
                    "鉴权接口-首页设置方法"
                ],
                "summary": "上传头像",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "上传头像",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "define.AddFoodRequest": {
            "type": "object",
            "properties": {
                "foodicon": {
                    "type": "string"
                },
                "foodname": {
                    "type": "string"
                },
                "foodprocedure": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "remarks": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "video": {
                    "type": "string"
                }
            }
        },
...
        "define.PwdDataType": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                }
            }
        },
        "define.UpdateFoodRequest": {
            "type": "object",
            "properties": {
                "foodicon": {
                    "type": "string"
                },
                "foodname": {
                    "type": "string"
                },
                "foodprocedure": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "remarks": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "video": {
                    "type": "string"
                }
            }
        },
        "define.UpdateInfoType": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "sex": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}